Fix all (relevant) errors detected by cppcheck

+++ include/rhino/common.h
@@ -10,8 +10,8 @@
 
 #define KERNEL_VBASE 0xC0000000
 
-#define BIT_IS_SET(var, pos) ((var) & (1 << (pos)))
-#define BIT_IS_CLEAR(var, pos) (!((var) & (1 << (pos))))
+#define BIT_IS_SET(var, pos) (((uint64_t)var) & (1ULL << (pos)))
+#define BIT_IS_CLEAR(var, pos) (!(((uint64_t)var) & (1ULL << (pos))))
 #define BIT_SET(var, pos) ((var) |= (1ULL << (pos)))
 #define BIT_CLEAR(var, pos) ((var) &= ~(1ULL << (pos)))
